made the api safer

+++ Finance Tracker/Application/Banks/Commands/DeleteBankCommand.cs
@@ -11,6 +11,7 @@ namespace Application.Banks.Commands;
 public record DeleteBankCommand : IRequest<Result<Bank, BankException>>
 {
     public required Guid BankId { get; init; }
+    public required Guid UserIdFromToken { get; set; }
 }
 
 public class DeleteBankCommandHandler(IBankRepository bankRepository, IUserRepository userRepository)
@@ -21,25 +22,41 @@ public class DeleteBankCommandHandler(IBankRepository bankRepository, IUserRepos
     {
         var bankId = new BankId(request.BankId);
         var existingBank = await bankRepository.GetById(bankId, cancellationToken);
-        return await existingBank.Match<Task<Result<Bank, BankException>>>(async bank =>
-            {
-                var existingUser = await userRepository.GetById(bank.UserId, cancellationToken);
 
-                return await existingUser.Match<Task<Result<Bank, BankException>>>(
-                    async user => await DeleteEntity(user, bank, cancellationToken),
-                    () => Task.FromResult<Result<Bank, BankException>>(new UserNotFoundException(bank.UserId)));
+        return await existingBank.Match<Task<Result<Bank, BankException>>>(
+            async b =>
+            {
+                var userFromBankId = b.UserId;
+                var existingUserFromBank = await userRepository.GetById(userFromBankId, cancellationToken);
+
+                return await existingUserFromBank.Match(
+                    async ufb =>
+                    {
+                        var userIdFromToken = new UserId(request.UserIdFromToken);
+                        var existingUserFromToken = await userRepository.GetById(userIdFromToken, cancellationToken);
+
+                        return await existingUserFromToken.Match<Task<Result<Bank, BankException>>>(
+                            async uft => await DeleteEntity(ufb,uft, b, cancellationToken),
+                            () => Task.FromResult<Result<Bank, BankException>>(new UserNotFoundException(b.UserId)));
+                    },
+                    () => Task.FromResult<Result<Bank, BankException>>(new UserNotFoundException(userFromBankId)));
             },
             () => Task.FromResult<Result<Bank, BankException>>(new BankNotFoundException(bankId)));
     }
 
-    private async Task<Result<Bank, BankException>> DeleteEntity(User user, Bank bank,
+    private async Task<Result<Bank, BankException>> DeleteEntity(User userFromBank,User userFromToken, Bank bank,
         CancellationToken cancellationToken)
     {
         try
         {
-            user.AddToBalance(bank.Balance);
-            await userRepository.Update(user, cancellationToken);
-            return await bankRepository.Delete(bank, cancellationToken);
+            if (userFromToken.Id == userFromBank.Id || userFromToken.IsAdmin)
+            {
+                userFromBank.AddToBalance(bank.Balance);
+                await userRepository.Update(userFromBank, cancellationToken);
+                return await bankRepository.Delete(bank, cancellationToken);
+            }
+
+            return await Task.FromResult<Result<Bank, BankException>>(new YouDoNotHaveTheAuthorityToDo(userFromToken.Id, userFromBank.Id));
         }
         catch (Exception exception)
         {
